+++ src/lib.rs
@@ -77,7 +77,9 @@
 //!
 //! Use [reduce()] to apply inter-bitset operation between elements of bitsets iterator.
 //!
-//! Use [apply()]  to apply inter-bitset operation between two bitsets. Also [&], [|], [`^`], [-].
+//! Use [apply()]  to apply inter-bitset operation between two bitsets.
+//! Also [&], [|], [`^`], [-] over &[BitSet], or corresponding [BitSetInterface]
+//! members.
 //!
 //! You can define your own inter-bitset operation by implementing [BitSetOp].
 //!
@@ -90,6 +92,16 @@
 //!
 //! Use `BitSet::from(impl BitSetInterface)` instead of collecting iterator for
 //! faster materialization into BitSet.
+//!
+//! # Eager inter-bitset operations
+//!
+//! Thou we encourage to use lazy inter-bitset operations whenever possible,
+//! there are optimized eager in-place operations in [BitSet]:
+//! * [`unite()`] and it's operator counterpart `a |= &b`.
+//! * [`into_union()`] and it's operator counterpart `a | b`.
+//!
+//! [`unite()`]: BitSet::unite
+//! [`into_union()`]: BitSet::into_union
 //!
 //! # Cursor
 //!